Apply improved types & architecture to Algorand
