+++ libs/coin-modules/coin-algorand/src/api/indexer.types.ts
@@ -15,27 +15,27 @@ export interface AlgoTransaction {
   senderAddress: string; // sender: 'RGX5XA7DWZOZ5SLG4WQSNIFKIG4CNX4VOH23YCEX56523DQEAL3QL56XZM',
   senderRewards: BigNumber; // 'sender-rewards': 0,
   recipientRewards: BigNumber; //'receiver-rewards': 0,
-  closeRewards?: BigNumber; // 'close-rewards': 0,
-  closeAmount?: BigNumber; //'closing-amount': 0,
+  closeRewards: BigNumber | undefined; // 'close-rewards': 0,
+  closeAmount: BigNumber | undefined; //'closing-amount': 0,
   fee: BigNumber;
   note: string;
 
   type: string; // 'tx-type': 'pay'
-  details?: AlgoTransactionDetails;
+  details: AlgoTransactionDetails | undefined;
 }
 
 export interface AlgoPaymentInfo {
   amount: BigNumber;
   recipientAddress: string; // receiver
-  closeAmount?: BigNumber; // close-amount
-  closeToAddress?: string; // close-remainder-to
+  closeAmount: BigNumber | undefined; // close-amount
+  closeToAddress: string | undefined; // close-remainder-to
 }
 
 export interface AlgoAssetTransferInfo {
   assetId: string; // asset-id
   assetAmount: BigNumber; // amount
   assetRecipientAddress: string; // receiver
-  assetSenderAddress?: string; // sender
-  assetCloseAmount?: BigNumber; // close-amount
-  assetCloseToAddress?: string; // close-to
+  assetSenderAddress: string | undefined; // sender
+  assetCloseAmount: BigNumber | undefined; // close-amount
+  assetCloseToAddress?: string | undefined; // close-to
 }
