+++ libs/coin-modules/coin-algorand/src/estimateMaxSpendable.ts
@@ -1,0 +1,46 @@
+import { getMainAccount } from "@ledgerhq/coin-framework/account/index";
+import { getAbandonSeedAddress } from "@ledgerhq/cryptoassets";
+import type { AccountBridge } from "@ledgerhq/types-live";
+import { BigNumber } from "bignumber.js";
+import createTransaction from "./createTransaction";
+import { getEstimatedFees } from "./getFeesForTransaction";
+import { computeAlgoMaxSpendable } from "./logic";
+import type { AlgorandAccount, AlgorandTransaction } from "./types";
+
+export const estimateMaxSpendable: AccountBridge<AlgorandTransaction>["estimateMaxSpendable"] =
+  async ({ account, parentAccount, transaction }) => {
+    const mainAccount = getMainAccount(account, parentAccount);
+    const { algorandResources } = mainAccount as AlgorandAccount;
+    if (!algorandResources) {
+      throw new Error("Algorand account expected");
+    }
+
+    const tx = {
+      ...createTransaction(account),
+      subAccountId: account.type === "Account" ? null : account.id,
+      ...transaction,
+      recipient: transaction?.recipient || getAbandonSeedAddress(mainAccount.currency.id),
+      useAllAmount: true,
+    };
+
+    const tokenAccount =
+      tx.subAccountId &&
+      mainAccount.subAccounts &&
+      mainAccount.subAccounts.find(ta => ta.id === tx.subAccountId);
+
+    if (tokenAccount) {
+      return tokenAccount.balance;
+    } else {
+      const fees = await getEstimatedFees(mainAccount, tx);
+
+      let maxSpendable = computeAlgoMaxSpendable({
+        accountBalance: mainAccount.balance,
+        nbAccountAssets: algorandResources.nbAssets,
+        mode: tx.mode,
+      });
+
+      maxSpendable = maxSpendable.minus(fees);
+
+      return maxSpendable.gte(0) ? maxSpendable : new BigNumber(0);
+    }
+  };
